api/pyproject.toml
[project]
name = "portfolium-api"
version = "0.3.0"
description = "Portfolium investment tracking API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pydantic[email]>=2.5.3",
    "yfinance>=0.2.36",
    "python-multipart>=0.0.6",
    "pandas>=2.2.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "scalar-fastapi>=1.4.0",
    "pillow>=10.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.8.0",
    "weasyprint>=66.0",
    "alembic>=1.13.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.0",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "flower>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.1.0",
    "pytest-mock>=3.12.0",
    "faker>=20.0.0",
    "factory-boy>=3.3.0",
    "black>=24.1.1",
    "ruff>=0.1.14",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["logs*", "tests*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
